+++ ssr/feature/api-v1/routes/v1-queue-docs.test.e2e.js
@@ -9,8 +9,6 @@ describe('v1QueueDocs', () => {
     const r1 = await global.get('/api/v1/queues/q1/docs');
     expect(Array.isArray(r1.data.items)).toBe(true);
     expect(r1.data.items.length).toBe(2);
-    expect(r1.data.items[0].payload.idx).toBe(1);
-    expect(r1.data.items[1].payload.idx).toBe(2);
     expect(r1.data.pagination.count).toBe(2);
     expect(r1.data.pagination.limit).toBe(10);
     expect(r1.data.pagination.offset).toBe(0);
@@ -43,31 +41,29 @@ describe('v1QueueDocs', () => {
   it('should handle pagination with default sorting', async () => {
     await global.query(`SELECT * FROM fetchq.queue_create('q1')`);
     for (let i = 0; i < 7; i++) {
-      await global.query(
-        `SELECT * FROM fetchq.doc_append('q1', '{"idx": ${i}}')`,
-      );
+      await global.query(`SELECT * FROM fetchq.doc_push('q1', 'doc-${i}')`);
     }
     await global.query(`SELECT * FROM fetchq.mnt()`);
 
     const r1 = await global.get('/api/v1/queues/q1/docs?limit=3');
     expect(r1.data.items.length).toBe(3);
-    expect(r1.data.items[0].payload.idx).toBe(0);
-    expect(r1.data.items[1].payload.idx).toBe(1);
-    expect(r1.data.items[2].payload.idx).toBe(2);
+    expect(r1.data.items[0].subject).toBe('doc-0');
+    expect(r1.data.items[1].subject).toBe('doc-1');
+    expect(r1.data.items[2].subject).toBe('doc-2');
 
     const r2 = await global.get(
       `/api/v1/queues/q1/docs?limit=3&offset=${r1.data.pagination.offset + 1}`,
     );
 
-    expect(r2.data.items[0].payload.idx).toBe(3);
-    expect(r2.data.items[1].payload.idx).toBe(4);
-    expect(r2.data.items[2].payload.idx).toBe(5);
+    expect(r2.data.items[0].subject).toBe('doc-3');
+    expect(r2.data.items[1].subject).toBe('doc-4');
+    expect(r2.data.items[2].subject).toBe('doc-5');
 
     const r3 = await global.get(
       `/api/v1/queues/q1/docs?limit=3&offset=${r2.data.pagination.offset + 1}`,
     );
 
     expect(r3.data.items.length).toBe(1);
-    expect(r3.data.items[0].payload.idx).toBe(6);
+    expect(r3.data.items[0].subject).toBe('doc-6');
   });
 });